minor edit: moved the `getAndSetAnonymousIdFromLocalStorage` function from `url_utils.ts` to `rudderstack_initialize.ts` file because it is not a URL related utility.
